\name{f_string}
\alias{f_string}
\title{Make a string of fourier functions}
\usage{
  f_string(n)
}
\arguments{
  \item{n}{order of the string}
}
\value{
  A string of fourier functions of length n
}
\description{
  Make a string of fourier functions
}
\note{
  a bit specialized to the Climarkplus
}

