\name{make_bernoulli_vector}
\alias{make_bernoulli_vector}
\title{change a probability and number or tries
to success/failure form}
\usage{
  make_bernoulli_vector(probs, atts)
}
\arguments{
  \item{probs}{Probability of success}

  \item{atts}{number of attempts}
}
\description{
  change a probability and number or tries to
  success/failure form
}

