\name{make_eval_other_string}
\alias{make_eval_other_string}
\title{Create a string that can be evalutaed to give
the offset from the "others" part of the fit}
\usage{
  make_eval_other_string(cs)
}
\arguments{
  \item{cs}{the coeficients of "others" parts of the fit}
}
\description{
  Create a string that can be evalutaed to give the offset
  from the "others" part of the fit
}
\details{
  This function uses R string manipulation to produce a
  string that can be evaluated as R code
}

