\name{make_markov_string}
\alias{make_markov_string}
\title{A simple utility to make the fourier part
of the fit string}
\usage{
  make_markov_string(params, levels, is_rain = FALSE)
}
\arguments{
  \item{params}{a list of parameters}

  \item{levels}{the levels to include in the fit}

  \item{is_rain}{names slightly different if called for
  amount of rain fit}
}
\value{
  a part of the fit string
}
\description{
  A simple utility to make the fourier part of the fit
  string
}

