\name{write_pl}
\alias{write_pl}
\title{A simple utility to write out a parameter
file}
\usage{
  write_pl(params, filename)
}
\arguments{
  \item{params}{a vector of parameters}

  \item{filname}{the filename to write to}
}
\description{
  A simple utility to write out a parameter file
}
\note{
  No error checking is done
}

